03_sqlite_to_postgres/requirements.txt
psycopg2==2.9.3
flake8==4.0.1
python-dotenv==0.19.2